add files for components such as materials and lights. Solve problem where the ray dir was not updating. Was used during debugging equal indecies of refraction and was not removed
