+++ a3/raytracerLinux/Colour.h
@@ -1,0 +1,24 @@
+#pragma once
+
+#include <ostream>
+
+class Colour {
+public:
+	Colour();
+	Colour(double r, double g, double b);
+	Colour(const Colour& other);
+
+	Colour& operator =(const Colour& other);
+	Colour operator *(const Colour& other);
+	double& operator[](int i);
+	double operator[](int i) const;
+
+	void clamp();
+
+private:
+	double m_data[3];
+};
+
+Colour operator *(double s, const Colour& c);
+Colour operator +(const Colour& u, const Colour& v);
+std::ostream& operator <<(std::ostream& o, const Colour& c);
