+++ a3/raytracerLinux/Euclidean.h
@@ -1,0 +1,92 @@
+#pragma once
+
+#include <iostream>
+
+class Point3D {
+public:
+	Point3D(); 
+	Point3D(double x, double y, double z);  
+	Point3D(const Point3D& other); 
+	bool isClose(const Point3D& other);
+	Point3D& operator =(const Point3D& other); 
+	double& operator[](int i); 
+	double operator[](int i) const; 
+
+private:
+	double m_data[3];
+};
+
+class Vector3D {
+public:
+	Vector3D(); 
+	Vector3D(double x, double y, double z); 
+	Vector3D(const Vector3D& other); 
+
+	Vector3D& operator =(const Vector3D& other); 
+	double& operator[](int i);  
+	double operator[](int i) const;  
+
+	double length() const; 
+	double normalize();
+	double dot(const Vector3D& other) const; 
+	Vector3D cross(const Vector3D& other) const; 
+
+    bool isNormalized() const;
+
+private:
+	double m_data[3];
+};
+
+// standard operators on points and vectors
+Vector3D operator *(double s, const Vector3D& v); 
+Vector3D operator +(const Vector3D& u, const Vector3D& v); 
+Point3D operator +(const Point3D& u, const Vector3D& v); 
+Vector3D operator -(const Point3D& u, const Point3D& v); 
+Vector3D operator -(const Vector3D& u, const Vector3D& v); 
+Vector3D operator -(const Vector3D& u); 
+Point3D operator -(const Point3D& u, const Vector3D& v); 
+Vector3D cross(const Vector3D& u, const Vector3D& v); 
+std::ostream& operator <<(std::ostream& o, const Point3D& p); 
+std::ostream& operator <<(std::ostream& o, const Vector3D& v); 
+
+class Vector4D {
+public:
+	Vector4D(); 
+	Vector4D(double w, double x, double y, double z); 
+	Vector4D(const Vector4D& other); 
+
+	Vector4D& operator =(const Vector4D& other); 
+	double& operator[](int i);  
+	double operator[](int i) const;  
+
+private:
+	double m_data[4];
+};
+
+class Matrix4x4 {
+public:
+  Matrix4x4(); 
+  Matrix4x4(const Matrix4x4& other); 
+  Matrix4x4& operator=(const Matrix4x4& other); 
+
+  Vector4D getRow(int row) const; 
+  double *getRow(int row); 
+  Vector4D getColumn(int col) const; 
+
+  Vector4D operator[](int row) const; 
+  double *operator[](int row); 
+
+  Matrix4x4 transpose() const; 
+		
+private:
+  double m_data[16];
+};
+
+Matrix4x4 operator *(const Matrix4x4& M, const Matrix4x4& N); 
+Vector3D operator *(const Matrix4x4& M, const Vector3D& v); 
+Point3D operator *(const Matrix4x4& M, const Point3D& p);
+// Multiply n by the transpose of M, useful for transforming normals.  
+// Recall that normals should be transformed by the inverse transpose 
+// of the matrix.  
+Vector3D transNorm(const Matrix4x4& M, const Vector3D& n); 
+std::ostream& operator <<(std::ostream& os, const Matrix4x4& M);
