+++ a3/raytracerLinux/Intersection.cpp
@@ -1,0 +1,7 @@
+#include "Intersection.h"
+
+Intersection::Intersection() :
+	enteringMaterial(nullptr),
+	t_value(0),
+	none(true)
+{}
