+++ a3/raytracerLinux/raytracer.h
@@ -1,5 +1,4 @@
 #pragma once
-#include "util.h"
 #include "scene_object.h"
 #include "light_source.h"
