--- a3/raytracerW32/Common.h
@@ -1,7 +1,0 @@
-#pragma once
-
-#define M_PI	3.14159265358979323846
-
-#define EPSILON	0.00001
-
-#define RECURSIVE_RECURRANCE 2
